+++ spec/support/ccms_environment_vars.rb
@@ -1,0 +1,11 @@
+module CCMS
+  def modified_environment_vars
+    {
+      'SOAP_CLIENT_USERNAME' => 'my_soap_client_username',
+      'SOAP_CLIENT_PASSWORD' => 'xxxxx',
+      'SOAP_CLIENT_PASSWORD_TYPE' => 'password_type',
+      'USER_LOGIN' => 'my_login',
+      'USER_ROLE' => 'my_role'
+    }
+  end
+end
